Refactor: modify module package struct
모듈 패키지 구조 변경 및 Querydsl 추가
