+++ module-core/module-jpa/build.gradle
@@ -1,3 +1,10 @@
 dependencies {
     api 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    api 'com.querydsl:querydsl-jpa'
+    api 'com.querydsl:querydsl-core'
+
+    annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jpa"
+    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
 }
