+++ module-core/module-jpa/src/main/java/com/example/modulejpa/coupon/domain/Coupon.java
@@ -1,0 +1,16 @@
+package com.example.modulejpa.coupon.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Coupon {
+
+    private Long id;
+    private Long memberId;
+}
