+++ module-core/module-jpa/src/main/java/com/example/modulejpa/coupon/entitiy/CouponEntity.java
@@ -1,0 +1,22 @@
+package com.example.modulejpa.coupon.entitiy;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+@Table(name = "coupon")
+public class CouponEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Long memberId;
+}
