--- module-core/module-jpa/src/main/java/com/example/modulejpa/domain/Member.java
@@ -1,18 +1,0 @@
-package com.example.modulejpa.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Member {
-
-    private Long id;
-    private String name;
-    private String email;
-    private String phoneNumber;
-}
