--- module-core/module-jpa/src/main/java/com/example/modulejpa/entity/CouponEntity.java
@@ -1,22 +1,0 @@
-package com.example.modulejpa.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-@Table(name = "coupon")
-public class CouponEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long memberId;
-}
