+++ module-core/module-jpa/src/main/java/com/example/modulejpa/member/entitiy/MemberEntity.java
@@ -1,0 +1,24 @@
+package com.example.modulejpa.member.entitiy;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Builder
+@Table(name = "member")
+public class MemberEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private String email;
+    private String phoneNumber;
+}
