+++ module-core/module-jpa/src/main/java/com/example/modulejpa/member/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.example.modulejpa.member.repository;
+
+import com.example.modulejpa.member.entitiy.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+}
