--- module-core/module-jpa/src/main/java/com/example/modulejpa/repository/CouponJpaRepository.java
@@ -1,9 +1,0 @@
-package com.example.modulejpa.repository;
-
-import com.example.modulejpa.entity.CouponEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CouponJpaRepository extends JpaRepository<CouponEntity, Long> {
-}
