--- module-core/module-jpa/src/main/java/com/example/modulejpa/repository/MemberJpaRepository.java
@@ -1,9 +1,0 @@
-package com.example.modulejpa.repository;
-
-import com.example.modulejpa.entity.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberJpaRepository extends JpaRepository<MemberEntity, Long> {
-}
