+++ module-producer/src/main/java/com/example/moduleproducer/coupon/service/CouponService.java
@@ -1,0 +1,16 @@
+package com.example.moduleproducer.coupon.service;
+
+import com.example.moduleredis.repository.CouponCountRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CouponService {
+
+    private final CouponCountRepository couponCountRepository;
+
+    public Long getNumberOfCouponIssued() {
+        return couponCountRepository.increment();
+    }
+}
