--- module-producer/src/main/java/com/example/moduleproducer/request/CreateMemberRequest.java
@@ -1,17 +1,0 @@
-package com.example.moduleproducer.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CreateMemberRequest {
-
-    private String name;
-    private String email;
-    private String phoneNumber;
-}
