+++ module-producer/src/test/java/com/example/moduleproducer/service/ApplyServiceTest.java
@@ -1,36 +1,8 @@
 package com.example.moduleproducer.service;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 @SpringBootTest
 class ApplyServiceTest {
 
-    @Autowired
-    private ApplyService applyService;
-
-    @Test
-    public void manyRequestByUniqueUser() throws Exception {
-        int threadCount = 1000;
-
-        ExecutorService executorService = Executors.newFixedThreadPool(32);
-        CountDownLatch countDownLatch = new CountDownLatch(threadCount);
-
-        for (int i = 0; i < threadCount; i++) {
-            executorService.submit(() -> {
-                try {
-                    applyService.apply(1L);
-                } finally {
-                    countDownLatch.countDown();
-                }
-            });
-        }
-
-        countDownLatch.await();
-    }
 }
